Split M4 pair setup step into stages
SRP PMS generation is very expensive on ESP8266 and takes about 5
seconds. With non-preemptive SDK this leads to device missing enough
beacons to get dropped by the AP and pairing process failing.
This happens often enough to be annoying

With this change SRP PMS generation happens in stages, allowing other
stuff to get done. The longest step is 2, taking 2.7 seconds, and there
are still occasional connection drops, but the rate of success overall
has improved a lot.

https://github.com/mongoose-os-apps/shelly-homekit/issues/748